- Full restructured - Full reworked
